Use open file description locks on Linux
